+++ socio-service/src/main/java/com/gym/socio/service/SocioService.java
@@ -20,19 +20,27 @@ public class SocioService {
         return socioRepository.findAll().stream().map(this::convertirDto).toList();
     }
 
-    public SocioDTO buscarPorId(Integer id){
-        Socio soci =socioRepository.findById(id).orElseThrow(() -> new RuntimeException("¡Usuario no encontrado!"));
-        return convertirDto(soci);
+    public SocioDTO buscarPorId(Integer id) {
+    if (id == null) {//validamos porque findbyid espera un valor que no sea nulo 
+        throw new RuntimeException("El ID del socio es obligatorio");
+    }
+    Socio soci = socioRepository.findById(id)
+            .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+    return convertirDto(soci);
     }
 
+
     public String eliminar(Integer id){
+    if (id == null) {
+        return "El ID del socio es obligatorio";
+    }    
     try {
             Socio socio = socioRepository.findById(id)
                     .orElseThrow(() -> new RuntimeException("¡Imposible eliminar! El Usuario con ID " + id + " no existe."));
-           socioRepository.delete(socio);
-            return "El Usuario '" + socio.getNombre() + "' ha sido retirado del registro exitosamente.";
-    } catch (RuntimeException e) {
-            return e.getMessage();
+            socioRepository.delete(socio);
+                return "El Usuario '" + socio.getNombre() + "' ha sido retirado del registro exitosamente.";
+        } catch (RuntimeException e) {
+                return e.getMessage();
         }
     }
     
@@ -41,7 +49,13 @@ public class SocioService {
        return socioRepository.save(socio);
     }
 
+
     public Socio actualizarSocio(Integer id,Socio socio){
+
+        if (id == null) {
+        throw new RuntimeException("El ID del socio es obligatorio");
+        }
+
         Socio soci = socioRepository.findById(id).orElseThrow(() -> new RuntimeException("¡El socio no existe en los registros!"));
         if(socio.getRut() != null){
             soci.setRut(socio.getRut());
